mv repo and pr fixt to parent tests/conftest.py
